% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_methods.R
\name{glogFit}
\alias{glogFit}
\title{Glog fit}
\usage{
glogFit(x, a = 1, inverse = FALSE)
}
\arguments{
\item{x}{Values to convert.}

\item{a}{glog offset.}

\item{inverse}{To apply glog or reverse back to original values.}
}
\description{
Glog fit
}
