% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_methods.R
\name{logFit}
\alias{logFit}
\title{Transform data using natural logarithm}
\usage{
logFit(x, a = 1, inverse = FALSE)
}
\arguments{
\item{x}{Values to convert.}

\item{a}{log offset.}

\item{inverse}{To apply log or reverse back to original values.}
}
\description{
Transform data using natural logarithm
}
