% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corrected_data.R
\name{plot_sbc}
\alias{plot_sbc}
\title{Plot the output from signal batch correction for the first 100 features}
\usage{
plot_sbc(df, corrected_df, classes, batch, order,
  output = "sbcms_plots.pdf")
}
\arguments{
\item{df}{A data frame containing the original data before correction (samples in columns, features in rows).}

\item{corrected_df}{A data frame containing the corrected data, as output by QCRSC}

\item{classes}{A factor or character vector of sample classes. All QC samples should be labelled "QC".}

\item{batch}{A vector indicating the batch each sample was measured in. If only one batch was measured then all values should be set to 1}

\item{order}{A numeric vector  indicating the order in which samples were measured.}

\item{output}{Filename of the output pdf file. Can include the path.}
}
\description{
Plot the output from signal batch correction for the first 100 features
}
