% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sbcdata}
\alias{sbcdata}
\title{Direct-infusion mass spectrometry (DIMS) data set}
\format{A list of 3 elements}
\source{
https://www.ebi.ac.uk/metabolights/MTBLS79
}
\usage{
sbcdata
}
\description{
Data set of 20 biological samples that were analysed repeatedly,
in 8 batches across 7 days.
}
\references{
Kirwan et al, Scientific Data volume 1,
Article number: 140012 (2014) https://www.nature.com/articles/sdata201412
}
\keyword{datasets}
