% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testData.R
\docType{data}
\name{testData}
\alias{testData}
\title{Testing data set}
\format{A list of outputs of sbcms functions}
\source{
https://www.ebi.ac.uk/metabolights/MTBLS79
}
\usage{
testData
}
\description{
Data for testthat tests
}
\keyword{datasets}
